Update set-up-checklist.md
Deleted Global Administrator as the permissions required for apps registration. I've seen multiple cases of GA being requested (although the apps were already registered) just because it is in the checklist. It takes a lot of effort from Identity team to explain why GA cannot and should not be granted.